+++ src/nextjs-front/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "next": "12.0.8",
     "react": "17.0.2",
+    "react-awesome-reveal": "^3.8.1",
     "react-dom": "17.0.2",
     "react-icons": "^4.3.1",
     "sharp": "^0.29.3"
