+++ src/nextjs-front/pages/index.tsx
@@ -1,72 +1,240 @@
-import type { NextPage } from 'next'
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
+import withWildLayout from "../components/hoc/withWildLayout";
+import Link from "next/link";
+import Image from "next/image";
+import { TiArrowDown } from "react-icons/ti";
+import { Feature, features } from "../constants/feature";
+import { team, TeamMember } from "../constants/team";
+import { VscGithub } from "react-icons/vsc";
+import { SiLinkedin } from "react-icons/si";
+import { Fade } from "react-awesome-reveal";
 
-const Home: NextPage = () => {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+const Feature: React.FC<Feature> = ({ label, description, Icon }) => (
+  <div className="flex flex-col items-center justify-between h-full text-xl text-center text-neutral-200 gap-y-8">
+    <div className="flex flex-col items-center gap-8">
+      <div className="flex items-center justify-center p-5 text-white bg-pink-600 rounded-full drop-shadow-md">
+        <Icon className="text-6xl fill-white" />
+      </div>
+      <Fade direction="down" duration={800}>
+        <h3 className="text-3xl font-bold text-white">{label}</h3>
+      </Fade>
     </div>
-  )
+    <Fade>
+      <p>{description}</p>
+    </Fade>
+  </div>
+);
+
+interface TeamMemberProps extends TeamMember {
+  fadeDirection?: "left" | "right";
 }
 
-export default Home
+const TeamMember: React.FC<TeamMemberProps> = ({
+  login42,
+  role,
+  firstname,
+  lastname,
+  imageSrc,
+  githubLink,
+  linkedinLink,
+  fadeDirection,
+}) => (
+  <Fade direction={fadeDirection} duration={800}>
+    <div className="relative z-10 h-96 drop-shadow-md">
+      <Image
+        alt={`${firstname} ${lastname}'s picture`}
+        src={imageSrc}
+        layout="fill"
+        objectFit="cover"
+        objectPosition="center"
+      ></Image>
+
+      <div className="absolute inset-0 z-20 flex flex-col items-center justify-between h-full p-4 text-center text-gray-900 bg-gray-100 opacity-0 transition hover:opacity-90 gap-y-2 ">
+        <div className="flex flex-col items-center gap-y-2">
+          <div className="flex items-center gap-x-2">
+            <h4 className="text-lg font-bold">
+              {firstname} {lastname}
+            </h4>
+            <small className="gray-900">{login42}</small>
+          </div>
+          <h5 className="uppercase">{role}</h5>
+        </div>
+
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum
+          non pellentesque elit. Nam non eros sed elit consequat efficitur.
+        </p>
+
+        <div className="flex text-4xl gap-x-4">
+          {githubLink && (
+            <a
+              href={githubLink}
+              target="_blank"
+              rel="noreferrer"
+              className="hover:opacity-75 transition"
+            >
+              <VscGithub />
+            </a>
+          )}
+          {linkedinLink && (
+            <a
+              href={linkedinLink}
+              target="_blank"
+              rel="noreferrer"
+              className="hover:opacity-75 transition"
+            >
+              <SiLinkedin />
+            </a>
+          )}
+        </div>
+      </div>
+    </div>
+  </Fade>
+);
+
+const HomePage: React.FC = () => {
+  return (
+    <main>
+      {/* hero banner section */}
+      <section
+        className="bg-fixed bg-gray-900 bg-center bg-no-repeat bg-cover"
+        style={{
+          backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.90), rgba(0, 0, 0, 0.90)),
+      url('/landing.webp')`,
+        }}
+      >
+        {/* Hero banner */}
+        <div className="absolute z-0 hidden lg:block top-36 left-24 animate-blinkSlow">
+          <Image
+            src="/starplanet.webp"
+            alt="decorative planet"
+            height={400}
+            width={400}
+          />
+        </div>
+
+        <div className="absolute z-0 hidden lg:block bottom-24 right-32 animate-fuse">
+          <Image
+            src="/rocket.webp"
+            alt="decorative rocket"
+            height={400}
+            width={400}
+          />
+        </div>
+
+        <div
+          className="container relative flex flex-col items-center justify-center px-2 mx-auto text-center text-white sm gap-y-8"
+          style={{ minHeight: "95vh" }}
+        >
+          <h1 className="text-3xl leading-tight text-pink-600 uppercase lg:text-5xl md:text-5xl">
+            It{"'"}s time for your
+            <br /> <b>self-transcendance</b>
+          </h1>
+          <h3 className="text-2xl lg:text-3xl text-neutral-200">
+            Let’s have the most entertaining pong game of your life.
+          </h3>
+          <Link href="/">
+            <a className="px-6 py-2 text-xl uppercase bg-pink-600 drop-shadow-md text-bold text-neutral-200">
+              Play now
+            </a>
+          </Link>
+        </div>
+      </section>
+
+      <section className={"relative py-20 md:py-32 bg-gray-900 "} id="discover">
+        <div className="absolute top-0 left-0 right-0 flex justify-center -translate-y-10">
+          <div className="flex items-start p-3 bg-white rounded-full">
+            <TiArrowDown className="block text-6xl drop-shadow-md animate-upAndDown" />
+          </div>
+        </div>
+        <div className="container px-3 mx-auto text-neutral-200">
+          <div className="grid lg:grid-cols-2 gap-x-16 gap-y-8">
+            <div className="flex flex-col items-start ">
+              <Fade direction="down" duration={800}>
+                <div>
+                  <h2 className="pb-4 text-4xl text-center lg:text-left">
+                    The pong reboot 2022 needs
+                  </h2>
+                  <p className="py-4 text-xl">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+                    Vestibulum non pellentesque elit. Nam non eros sed elit
+                    consequat efficitur. Suspendisse efficitur libero eget diam
+                    semper, et eleifend nisl vulputate.
+                  </p>
+                  <p className="py-4 text-xl">
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. In
+                    luctus, lorem non efficitur laoreet, ante arcu vehicula
+                    eros, quis rutrum ligula ligula eu purus. Vestibulum tempus
+                    orci ipsum, eu condimentum neque egestas non.
+                  </p>
+                  <p className="py-4 text-xl">
+                    Nulla pulvinar luctus orci, facilisis faucibus urna tempus
+                    volutpat. Sed tempor tortor orci, sit amet consequat tellus
+                    mattis quis. Mauris vitae consequat mauris, non pretium
+                    velit.
+                  </p>
+                </div>
+              </Fade>
+              <Link href="/">
+                <a className="px-10 py-2 mx-auto mt-4 text-xl font-bold uppercase bg-pink-600 lg:mx-0 drop-shadow-md text-bold text-neutral-200">
+                  Try it
+                </a>
+              </Link>
+            </div>
+            <Fade direction="right" duration={1000}>
+              <Image
+                src="/pong_mac.webp"
+                alt="pong overview"
+                width={1300}
+                height={1096}
+              />
+            </Fade>
+          </div>
+        </div>
+      </section>
+
+      <section id="features" className="bg-gray-900">
+        <div className="flex flex-col items-center py-16 md:mx-auto md:container gap-y-8">
+          <div className="relative grid md:grid-cols-3 gap-16">
+            <div className="absolute hidden h-1 bg-white rounded left-48 right-48 top-12 lg:block" />
+            {features.map((feature) => (
+              <Feature key={feature.label} {...feature} />
+            ))}
+          </div>
+          <Link href="/">
+            <a className="px-10 py-2 mx-auto mt-4 text-xl font-bold uppercase bg-pink-600 drop-shadow-md text-bold text-neutral-200">
+              Enter the fight
+            </a>
+          </Link>
+        </div>
+      </section>
+
+      <section
+        id="team"
+        className="pt-16 pb-16 bg-center bg-cover"
+        style={{
+          backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.90), rgba(0, 0, 0, 0.90)),
+      url('/42paris.webp')`,
+        }}
+      >
+        <div className="flex flex-col px-2 md:container md:mx-auto gap-y-16">
+          <div className="flex flex-col gap-y-2">
+            <h2 className="text-5xl font-bold text-center text-pink-600 uppercase">
+              The team
+            </h2>
+            <h3 className="text-3xl text-center text-white">
+              {" "}
+              They made it happen!
+            </h3>
+          </div>
+          <div className="grid grid-cols-5 grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5 gap-8">
+            {team.map((member, index, arr) => (
+              <TeamMember key={member.login42} {...member} fadeDirection={index === Math.floor(arr.length / 2) ? undefined : index < arr.length / 2 ? 'right' : 'left'  } />
+            ))}
+          </div>
+        </div>
+      </section>
+    </main>
+  );
+};
+
+export default withWildLayout(HomePage);
